tx: always reply when the majority replied during read
If the 4 replicas contain (0, 0, not_found, not_found), a majority for 'ok' could not be found and the previous code waited for MajDeny even if an answer could be given by the (0, 0, not_found) quorum. We could therefore never read the stored 0 value as any read returned not_found - now quorums with two values return 0.

Note: If a read returns not_found and a write is added to the transaction afterwards, it can still be committed which will (incorrectly) return {ok}. The problem is not in the commit but in the fact, that the client got not_found and made his decision based on that.
A future commit will fix this behaviour.

git-svn-id: https://scalaris.googlecode.com/svn/trunk@4319 71766da9-9752-0410-ab17-bfafa751c56a