Wiki on Scalaris: extend ScalarisReadRandomListEntryOp1 to be able to read from one bucket only or all of them
This depends on the optimisation, e.g. optimisations with write caches read the value from the read-key plus one of the write caches. getRandomArticle() tries a single bucket first and if the fails, it reads all of them, also using ScalarisReadRandomListEntryOp1 for this part.
If failNotFound is set and several buckets are read, the NotFoundException will only be thrown if all buckets were not found!

git-svn-id: https://scalaris.googlecode.com/svn/trunk@4386 71766da9-9752-0410-ab17-bfafa751c56a