new backend db implementation
this changes the db system that actually stores data on the nodes. this
simpler architecture makes it much easier to implement new backend (ets is the
default and example implementation). an alternative implementation just needs to
fullfill the backend_beh behaviour. change the DB macro in db_dht.erl
accordingly.
this also fixes a possible deadlock of snapshots during silde operations.
there exists a problem in the rrepair_SUITE which was introduced by these
changes but the problem seems to be with rrepair. will be fixed soon.

git-svn-id: https://scalaris.googlecode.com/svn/trunk@4947 71766da9-9752-0410-ab17-bfafa751c56a