rrepair: restore rr_recon and rr_resolve shutdown after rrepair shutdown
Linking to the rrepair_process is not enough! Either exit needs to be trapped or a monitor has to be used. Since bloom spawns rr_recon from rr_recon, a monitor to the rrepair process was chosen. However, we do not need to keep the reference and demonitor since the rr_recon and rr_resolve are shut down after receiving the DOWN message anyway which will cleanup the monitor automatically.

git-svn-id: https://scalaris.googlecode.com/svn/trunk@4963 71766da9-9752-0410-ab17-bfafa751c56a