mr: add erlang <-> js encoding and decoding
most importantly 2-tuples are treated as key-value pairs and encoded as objects
where {k, v} becomes o[k] == v;
a list of 2-tuples becomes one object with all keys as properties. in order to
avoid key collisions the (possibly multiple) values to one key are aggregated
into a list of values. to be consistent, the same happens in erlang. so before a
reduce phase (map phases get each KV-pair individually) all values corresponding
to one key are aggregated in to a list of values.

git-svn-id: https://scalaris.googlecode.com/svn/trunk@5488 71766da9-9752-0410-ab17-bfafa751c56a