rt_frt: reduce the number of messages sent on lookups
Prior to this change, FRT send a message to a lookup's issuer for each lookup. This change
reduces the amount of messages by only sending a message when the second node in the path
determines that the first hop was suboptimal and that the issuer should know the first
hops successor.

git-svn-id: https://scalaris.googlecode.com/svn/trunk@5581 71766da9-9752-0410-ab17-bfafa751c56a