rrepair, recon: subscribe to the rrepair process instead (not rr_recon)
rr_recon finishes and sends a message back, at the same time the failure detector gets the monitor's DOWN message and sends a crash on the prio channel which may overtake the success message from the rr_recon process
-> subscribe to the long-lived rrepair process instead

Since rr_recon is a gen_component, we assume that it does not crash unless the whole node crashes. We also assume that the component always reports back which is tested by the unit tests.

git-svn-id: https://scalaris.googlecode.com/svn/trunk@5854 71766da9-9752-0410-ab17-bfafa751c56a