scalarisctl: remove scheduler flags
Although binding the scheduler threads to cores can can have significant
advantages, the benchmarking of different flags did not result in clear
advantages for one combination. The results depend on the benchmark used, the
number of Erlang VMs per physical machine and the size of the ring.

Therefore no flags will be set by default, but the environment variable
governing the use of flags will be left present, to allow for easy
optimization in individual cases.